mcp_servers/knowledge_registry/requirements.txt
psycopg2-binary==2.9.9
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.12.5
